LayoutTests/media/network-no-source-const-shadow.html

<!DOCTYPE html>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script>
description("Test that NETWORK_NO_SOURCE is immutable at the appropriate times");

shouldBe("HTMLMediaElement.NETWORK_NO_SOURCE", "3");
shouldBe("HTMLMediaElement.prototype.NETWORK_NO_SOURCE", "3");

var element = document.createElement("audio");

shouldBe("element.NETWORK_NO_SOURCE", "3");

HTMLMediaElement.NETWORK_NO_SOURCE = 13;
HTMLMediaElement.prototype.NETWORK_NO_SOURCE = 14;
element.NETWORK_NO_SOURCE = 15;

shouldBe("HTMLMediaElement.NETWORK_NO_SOURCE", "3");
shouldBe("HTMLAudioElement.prototype.NETWORK_NO_SOURCE", "3");
shouldBe("element.NETWORK_NO_SOURCE", "15");
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
