LayoutTests/media/network-no-source-const-shadow-expected.txt

Test that NETWORK_NO_SOURCE is immutable at the appropriate times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS HTMLMediaElement.NETWORK_NO_SOURCE is 3
PASS HTMLMediaElement.prototype.NETWORK_NO_SOURCE is 3
PASS element.NETWORK_NO_SOURCE is 3
PASS HTMLMediaElement.NETWORK_NO_SOURCE is 3
PASS HTMLAudioElement.prototype.NETWORK_NO_SOURCE is 3
PASS element.NETWORK_NO_SOURCE is 15
PASS successfullyParsed is true

TEST COMPLETE
